Object Storage Server Improvements (#99)
* Update CMake to work with new build backend

We no longer need to hard code (well, still a little bit) the version
number in CMakeLists.txt now. When invoked by the python build backend,
the backend automatically define some variables for us. We use them.

Also do not build tests when we are not manually running CMake. This is
temporary, and will be changed once we get real unit tests.

Signed-off-by: gxu <georgexu420@gmail.com>

* Create a run_object_storage_server.py on root directory

This is for debugging when engineering on git version (not install
package).

Signed-off-by: gxu <georgexu420@gmail.com>

* Update naming convention following C++ Coding Guidance

- Non-free functions are now renamed to camelCase;
- Variables are now renamed to camelCase as well;
- Tests and free functions are not yet renamed, reasons as follows:
  - we will get rid of most tests soon enough as we will write UT;
  - free functions have a long tradition of being snake style, we should
  discuss this.

Signed-off-by: gxu <georgexu420@gmail.com>

* Decouple updateRecord and network logic

Signed-off-by: gxu <georgexu420@gmail.com>

* Add unit test that doesn't need standalone server running

Update CMake script to build this new example. Don't remove old examples
aka test_scaler_object for now. Planning to remove it after the code is
redesigned.

Update the examples. The new example file is much shorter, cleaner,
doesn't require users to look at asio details, and most importantly,
doesn't need standalone server to run.

Signed-off-by: gxu <georgexu420@gmail.com>

* Update linteryml

Signed-off-by: gxu <georgexu420@gmail.com>

* Automatically run unittest after build

Signed-off-by: gxu <georgexu420@gmail.com>

* Fix a stupid error: Do not reenter directory in the second command

Signed-off-by: gxu <georgexu420@gmail.com>

* Catch exception for example

Signed-off-by: gxu <georgexu420@gmail.com>

* Bypass the download process for easier testing

Signed-off-by: gxu <georgexu420@gmail.com>

* Add pandas as requirement.

Signed-off-by: gxu <georgexu420@gmail.com>

* Resolve advices

Signed-off-by: gxu <georgexu420@gmail.com>

* Make sure the downloaded data can be reached

Signed-off-by: gxu <georgexu420@gmail.com>

---------

Signed-off-by: gxu <georgexu420@gmail.com>
Co-authored-by: gxu <georgexu420@gmail.com>